tests/wpt/tests/html/semantics/forms/the-select-element/customizable-select/selectedcontent-nested.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="https://github.com/whatwg/html/issues/11880">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<select>
  <button>
    <selectedcontent id=main></selectedcontent>
  </button>
  <option>one</option>
  <option>two</option>
</select>

<script>
test(() => {
  const select = document.querySelector('select');
  const mainSelectedcontent = document.querySelector('selectedcontent');
  const button = document.querySelector('button');

  const parentSelectedcontent = document.createElement('selectedcontent');
  parentSelectedcontent.id = 'parent';
  parentSelectedcontent.textContent = 'parent';
  const childSelectedcontent = document.createElement('selectedcontent');
  childSelectedcontent.id = 'child';
  childSelectedcontent.textContent = 'child';

  parentSelectedcontent.appendChild(childSelectedcontent);
  button.appendChild(parentSelectedcontent);

  assert_equals(parentSelectedcontent.innerHTML, 'one',
    'parent selectedcontent should not get cleared after inserting.');
  assert_equals(childSelectedcontent.innerHTML, 'child',
    'child selectedcontent should not be modified after inserting.');
  assert_equals(childSelectedcontent.parentNode, null,
    'child selectedcontent should get disconnected after inserting.');

  parentSelectedcontent.appendChild(childSelectedcontent);
  select.value = 'two';
  assert_equals(parentSelectedcontent.innerHTML, 'two',
    'parent selectedcontent should get updated when selected option is changed.');
  assert_equals(childSelectedcontent.innerHTML, 'two',
    'child selectedcontent should get updated when selected option is changed.');
  assert_equals(childSelectedcontent.parentNode, null,
    'child selectedcontent should get disconnected when selected option is changed.');
}, 'Modification behavior when nesting selectedcontent elements.');
</script>
